+++ tests/test_utilities/test_in2csv.py
@@ -1,6 +1,7 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
 
+import os
 import sys
 
 import six
@@ -170,3 +171,35 @@ class TestIn2CSV(CSVKitTestCase, EmptyFileTests):
 
         self.assertEqual(next(output), 'not this one\n')
         self.assertEqual(next(output), 'data\n')
+
+    def test_convert_xls_with_write_sheets(self):
+        try:
+            self.assertConverted('xls', 'examples/sheets.xls', 'examples/testxls_converted.csv', ['--sheet', 'data', '--write-sheets', "ʤ,1"])
+            with open('examples/sheets_0.csv', 'r') as f:
+                with open('examples/testxls_converted.csv', 'r') as g:
+                    self.assertEqual(f.read(), g.read())
+            with open('examples/sheets_1.csv', 'r') as f:
+                with open('examples/testxls_unicode_converted.csv', 'r') as g:
+                    self.assertEqual(f.read(), g.read())
+            self.assertFalse(os.path.exists('examples/sheets_2.csv'))
+        finally:
+            for suffix in (0, 1):
+                path = 'examples/sheets_%d.csv' % suffix
+                if os.path.exists(path):
+                    os.remove(path)
+
+    def test_convert_xlsx_with_write_sheets(self):
+        try:
+            self.assertConverted('xlsx', 'examples/sheets.xlsx', 'examples/testxlsx_converted.csv', ['--sheet', 'data', '--write-sheets', "ʤ,1"])
+            with open('examples/sheets_0.csv', 'r') as f:
+                with open('examples/testxlsx_converted.csv', 'r') as g:
+                    self.assertEqual(f.read(), g.read())
+            with open('examples/sheets_1.csv', 'r') as f:
+                with open('examples/testxlsx_unicode_converted.csv', 'r') as g:
+                    self.assertEqual(f.read(), g.read())
+            self.assertFalse(os.path.exists('examples/sheets_2.csv'))
+        finally:
+            for suffix in (0, 1):
+                path = 'examples/sheets_%d.csv' % suffix
+                if os.path.exists(path):
+                    os.remove(path)